Updated CMakeLists to copy qt shared libraries into distro.

+++ src/antpack/cpp_src/developability_calcs/liability_search_tool.cpp
@@ -47,7 +47,7 @@ LiabilitySearchToolCpp::analyze_seq(std::string sequence,
                     "please report."));
     }
 
-    int sequence_length = sequence.length();
+    size_t sequence_length = sequence.length();
     std::vector<std::pair<std::pair<int, int>, std::string>> output;
 
     for (size_t i=0; i < sequence_length; i++) {
